requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
jpholiday>=0.1.8
ortools>=9.10.4067
gspread>=6.0.0
google-auth>=2.17.0
python-dateutil>=2.8.2
